pci-phb: Reimplement dma-map-in/out
The immediate problem with the code is that it relies on memory allocator
aligning addresses to the size. This is true for SLOF but not for GRUB
and in unaligned situations we end up mapping more pages than bm-alloc
allocated.

This fixes the problem by calculating aligned DMA size before calling
bm-alloc. While at this, simplify the code by removing global variables,
add tce-ps/mask and use them.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>